requirements.txt
requests>=2.31.0
supabase>=2.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
tqdm>=4.66.0
openai>=1.0.0
regex>=2023.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0 